Update and rename crm-card.js to example-card.js
